get_sentiment_stocktwits

This function computes the sentiment based on bullish and bearish tag from stocktwits using the last 30 twits.

get_sentiment_stocktwits(stock_symbol, path_twits, sentiment_index_type)

Arguments

stock_symbol
A vector with the stocks symbols.

path_twits
The path where the Json files will be stored.

sentiment_index_type
The sentiment type to be used according to the dictionary, positive, negative or both. Default is both, positive and negative

Value

A numeric value with the value of the sentiment index.

Examples

if (FALSE) {
#Not run:
path_twits <- 'your path'
  symbols <- c("EWZ", "SPX", "SPY", "USO")

 stocktwits_index <- get_sentiment_stocktwits(stock_symbol = symbols,
 path_twits = path_twits)
   }